refactor(contract): 이벤트 객체 패키지 이동
