+++ contract-service/src/main/java/com/example/contractservice/contract/service/dto/event/ContractEvent.java
@@ -1,4 +1,4 @@
-package com.example.contractservice.contract.service.event.dto;
+package com.example.contractservice.contract.service.dto.event;
 
 import java.time.Instant;
